modified:   css/global.css 	modified:   include/config.default.php 	modified:   include/geocoding_view.php 	modified:   include/header.php 	modified:   include/resource_functions.php 	modified:   include/searchbar.php 	modified:   index.php 	modified:   languages/en.php 	modified:   pages/collection_download.php 	modified:   pages/collection_request.php 	modified:   pages/collections.php 	modified:   pages/download.php 	modified:   pages/home.php 	modified:   pages/resource_request.php 	modified:   pages/resource_share.php 	new file:   pages/team/team_external_shares.php 	modified:   pages/terms.php 	new file:   pages/tools/test_coverage.php 	modified:   pages/user/user_preferences.php 	modified:   plugins/posixldapauth/posixldapauth.yaml 	modified:   plugins/rse_responsive/css/slim-style.css 	new file:   plugins/simpleldap/config/config.php 	new file:   plugins/simpleldap/dbstruct/table_simpleldap_groupmap.txt 	new file:   plugins/simpleldap/dbstruct/table_user.txt 	new file:   plugins/simpleldap/hooks/all.php 	new file:   plugins/simpleldap/hooks/team_user_edit.php 	new file:   plugins/simpleldap/hooks/user_password.php 	new file:   plugins/simpleldap/include/simpleldap_functions.php 	new file:   plugins/simpleldap/languages/en.php 	new file:   plugins/simpleldap/license.txt 	new file:   plugins/simpleldap/pages/ajax_test_auth.php 	new file:   plugins/simpleldap/pages/setup.php 	new file:   plugins/simpleldap/simpleldap.yaml 	modified:   tests/test.php 	new file:   tests/test_list/000800_escape_check.php

+++ tests/test.php
@@ -15,34 +15,103 @@ if (php_sapi_name()!=="cli") {exit("This utility is command line only.");}
 
   */
 
-  
-# Create a database for testing purposes
-$mysql_db="rs_test_db";
-echo "Creating database $mysql_db\n";ob_flush();
-sql_query("drop database if exists `$mysql_db`");
-sql_query("create database `$mysql_db`");
+$argv=preg_replace('/^(-|--|\/)/','',$argv);    // remove leading /, -- or -
 
-# Connect and create standard tables.
-echo "Creating default database tables...";ob_flush();
-sql_connect();
-check_db_structs(true);
-echo "...done\n";
+if(array_search('?',$argv)!==false || array_search('help',$argv)!==false)
+    {
+?>
+
+Command line paramaters:
+
+-nosetup        Do not setup the database, connect user in current state
+-noteardown     Do not drop the database once tests have completed
+-help or -?     This help information
+[n]...          Specific test number(s) to run
+<?php
+    exit;
+    }
+
+# Create an array of tests that were passed from the command line
+$specific_tests=array();
+foreach($argv as $arg)
+    {
+    if(is_numeric($arg))
+        {
+        array_push($specific_tests, str_pad($arg,6,'0',STR_PAD_LEFT));
+        }
+    }
+
+$mysql_db = "rs_test_db";
+$test_user_name = "admin";
+
+if(array_search('nosetup',$argv)===false)
+    {
+
+    # Create a database for testing purposes
+    echo "Creating database $mysql_db\n";
+    ob_flush();
+    sql_query("drop database if exists `$mysql_db`");
+    sql_query("create database `$mysql_db`");
+
+    # Connect and create standard tables.
+    echo "Creating default database tables...";
+    ob_flush();
+    sql_connect();
+    check_db_structs(true);
+    echo "...done\n";
+
+    # Insert a new user and run as them.
+    $u = new_user($test_user_name);
+    }
+else
+    {
+    sql_connect();
+
+    # Try to retrieve the ref of the existing user
+    $u = sql_value("SELECT `ref` AS value FROM `user` WHERE `username`='{$test_user_name}'",-1);
+    if ($u==-1)
+        {
+        die("Could not find existing '{$test_user_name}' user");
+        }
+    }
+
+# Setup user
+user_set_usergroup($u, 3);
+$userdata = get_user($u);
+setup_user($userdata);
+echo "Now running as user $userref\n";
+ob_flush();
 
 # Use an alternative filestore path
-$storagedir.="/rs_test/";
-if (!file_exists($storagedir)) {mkdir($storagedir);}
+$storagedir .= "/rs_test/";
+if (!file_exists($storagedir))
+    {
+    mkdir($storagedir);
+    }
 echo "Filestore is now at $storagedir\n";
 
-# Insert a new user and run as them.
-$u=new_user("admin");
-user_set_usergroup($u,3);
-$userdata=get_user($u);
-setup_user($userdata);
-echo "Now running as user $userref\n";ob_flush();
-
 # Get a list of tests
-$tests=scandir("test_list");
-$tests=array_filter($tests,function ($string) {return (strpos($string,".php")!==false);}); # PHP files only
+$tests = scandir("test_list");
+$tests = array_filter($tests, function ($string)
+    {
+    global $specific_tests;
+    if (strpos($string, ".php")===false)
+        {
+        return false;
+        }
+    if (count($specific_tests)==0)
+        {
+        return true;
+        }
+    foreach($specific_tests as $specific_test)
+        {
+        if (strpos($string, $specific_test)!==false)
+            {
+            return true;
+            }
+        }
+    return false;
+    }); # PHP files only
 asort($tests);
 
 # Run tests
@@ -60,5 +129,8 @@ foreach ($tests as $test)
     }
 echo "-----\nAll tests complete.\n";
 
-# Remove database
-sql_query("drop database `$mysql_db`");
+if(array_search('noteardown',$argv)===false)
+    {
+    # Remove database
+    sql_query("drop database `$mysql_db`");
+    }
